response tests for v3
